Update fr.json
Hello,
First of all, I would like to inform you that I have searched for your project at Transiflex, but without success.
Explanation of the change made and desired: 
The problem is indeed only a vocabulary problem and not a functional problem. In a form, when we talk about multiple or unique choices, we are talking about the possibilities of answers given to the people who will fill in the form. 
If the choice is multiple, it means that the person will be able to choose several answers (check boxes). 
If the choice is not multiple, and therefore unique, we are in the case of a radio box.
So, I think that the notion of "multiple" for the radio box misleads the users (whether it is the one who creates the form or the one who will fill it).
All this thinking is done in the French linguistic culture, so I don't question the translations in other languages.
Sincerely yours,
Jérôme