Feature: Allow shuffling of answer options
This introduces an question option to randomize the order of the answer options
by using the Fisher-Yates algorithm.
Shuffling the order of the answer choices reduces bias in responses.
Implements #1067

If no shuffling is enabled the options are sorted in the order they were
created, as currently they are shown as returned by the database which
is not necessarily sorted. Fixes #1007

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>