Add UI to preview and fill out own forms
This implements #1023 by providing a button to switch to the submit view
of an owned form like suggested in the [issue comment](https://github.com/nextcloud/forms/issues/1023#issuecomment-972117908).

This also adds the TopBar to the submit view (for owned forms)
to make the UX more consistent and changes the default view to the submit view.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>