Fix question header title for extra long text
Currently the title is set to a fixed height of 44px (used 16px text height + padding around)
to fit the menu button on the right. This change allows the header to have multiple lines of text
by setting only a minimum height + a fixed padding.

And also allows Question options to be longer than one line.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>