feat: Add technical identifiers for questions
Technical identifiers allow to map form questions to custom data.
On the frontend those identifiers are implemented as `name` props on the form elements.

The name attribute for technical identifier should be mapped to all options
so a `fieldset` is introduced around radio / checkbox buttons.
Added required aria label for `fieldset` (labeled by the question title).

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>