fix: Make sure "other" answers are correctly handled
In the backend it handling the "other" answer setting was not changed
after switching from object to array for extra settings.

In the frontend the value handling of the checkboxes or radio switches
was not correct as the initial values is always an empty array.
This lead to issues with radio switch answers.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>