fix(Submit): Make `answers` reactive and fix invalid mutation of computed property
Setting `answers` using `.sync` is not reactive as property changes are not triggering the computed / watch effect,
so instead use a `onUpdate` method to update the `answers` object in a way that will trigger any watch effect.

Also make `getFormValuesFromLocalStorage` a method instead a computed value.
Previously that computed value were assigned a new value which is invalid as computed values should be constant (except for computed setter but this was none).

And make sure "other" answer is set correctly so it will be saved in LocalStorage

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>